refactor server, tidy up, use handlers/middleware/utils pattern
