+++ src/server/handlers/themes-list-handler.tsx
@@ -1,0 +1,7 @@
+import type { Response } from "express";
+import { buildThemeList } from "../utils/build-themes-list";
+
+export default async ({ res }: { res: Response }) => {
+  res.type("json");
+  res.send(JSON.stringify(await buildThemeList()));
+};
